+++ src/main/java/DAO/Interfaces/CrudInterface.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface CrudInterface<T> {
     public void create(T obj);
-    public void update(T obj, int id);
-    public void delete(int id);
-    public T getById(int id) throws SQLException;
+    public void update(T obj, long id);
+    public void delete(long id);
+    public T getById(long id) throws SQLException;
     public List<T> getAll() throws SQLException;
 }
